shams_academy_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
bcrypt==4.0.1
celery==5.3.4
redis==5.0.1
Pillow==9.5.0
django-celery-results==2.5.1
django-celery-beat==2.5.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
django-cors-headers==4.3.1
python-magic==0.4.27
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9 